Fix sphere matrix transform
We were performing scale/translate in the wrong order, meaning
that if we changed the size of a sphere it would be moved closer
to the center and move at different speeds.

gl-matrix has a handy helper function for calculating a model
matrix based on an object's position, scale and rotation. We
could have continued to use scale/translate in the opposite order,
but mat4.fromRotationTranslationScale is both slightly faster
and less likely to pick up bugs like the above :)