Reduce the speed of gravity
Previously, the spheres were moving too quickly and falling
*through* the terrain. This can be mitigated by slowing them
down, and increasing their radius.

I think it might also be possible to improve this by increasing
the number of iterations in the physics solver, meaning that
they're less likely to miss a terrain collision, but I haven't
tried that out yet.